requirements.txt
Flask==2.0.3
stripe==2.56.0
pymongo==3.11.4
python-dotenv==0.17.1
cryptography==41.0.1